+++ CEPrivateMutableArray.h
@@ -41,8 +41,8 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 @interface CEPrivateMutableArray : NSMutableArray
 {
     id *_pointers;
-    unsigned int _size;
-    unsigned int _nPointers;
+    NSUInteger _size;
+    NSUInteger _nPointers;
 }
 
 @end
